+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
   - Allow to set advertised address when forcing a new Swarm cluster on a NuvlaBox
+  - Edge detail - Do not propose execution-mode in add or revoke SSH credentials modal
 
 ## [2.25.0] - 2021-10-21